refactor: Moved property classes into a more fitting organization where Group and Category are in their own package.
