+++ configuration/src/main/java/com/synopsys/integration/configuration/property/base/TypedProperty.kt
@@ -25,4 +25,14 @@ package com.synopsys.integration.configuration.property.base
 import com.synopsys.integration.configuration.parse.ValueParser
 import com.synopsys.integration.configuration.property.Property
 
+/**
+ * A property that is associated with a single type.
+ *
+ * It can't itself be retrieved from a Configuration but provides a shared parent class
+ * for Nullable and Valued properties, both of which need a property with a single type.
+ *
+ * @param T the type this property returns when it is retrieved from a Configuration.
+ * @property key a normalized lowercase dot separated key such as "example.key".
+ * @property parser the parser that converts string values to the given type T.
+ */
 abstract class TypedProperty<T>(key: String, val parser: ValueParser<T>) : Property(key) {}
