+++ configuration/src/main/java/com/synopsys/integration/configuration/source/MapPropertySource.kt
@@ -20,7 +20,30 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.configuration.property
+package com.synopsys.integration.configuration.source
 
-abstract class Category(val name: String) {
+import com.synopsys.integration.configuration.util.KeyUtils
+
+class MapPropertySource(private val givenName: String, private val underlyingMap: Map<String, String>) : PropertySource {
+    private val normalizedPropertyMap: Map<String, String> = underlyingMap.map { KeyUtils.normalizeKey(it.key) to it.value }.toMap()
+
+    override fun getOrigin(key: String): String? {
+        return givenName;
+    }
+
+    override fun hasKey(key: String): Boolean {
+        return normalizedPropertyMap.containsKey(key);
+    }
+
+    override fun getValue(key: String): String? {
+        return normalizedPropertyMap.getOrElse(key, { -> null });
+    }
+
+    override fun getName(): String {
+        return givenName;
+    }
+
+    override fun getKeys(): Set<String> {
+        return normalizedPropertyMap.keys
+    }
 }
