+++ configuration/src/test/java/com/synopsys/integration/configuration/config/MapPropertySourceTests.kt
@@ -1,5 +1,6 @@
 package com.synopsys.integration.configuration.config
 
+import com.synopsys.integration.configuration.source.MapPropertySource
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
